Moved distribution utils classes from checks package to utils package (#583)
* Moved distribution utils classes from checks package to utils package

* Return hashable to deepchecks type in train_test_feature_drift
