+++ deepchecks/checks/distribution/train_test_feature_drift.py
@@ -16,10 +16,10 @@ from typing import Union, List, Dict, Optional
 import pandas as pd
 
 from deepchecks import Dataset, CheckResult, TrainTestBaseCheck, ConditionResult
-from deepchecks.checks.distribution.dist_utils import calc_drift_and_plot
+from deepchecks.utils.distribution.drift import calc_drift_and_plot
 from deepchecks.utils.features import calculate_feature_importance_or_none
-from deepchecks.utils.typing import Hashable
 from deepchecks.errors import DeepchecksValueError
+from deepchecks.utils.typing import Hashable
 
 
 __all__ = ['TrainTestFeatureDrift']
