+++ deepchecks/checks/distribution/trust_score_comparison.py
@@ -13,9 +13,9 @@ import numpy as np
 import plotly.graph_objects as go
 
 from deepchecks import Dataset, CheckResult, TrainTestBaseCheck, ConditionResult, ConditionCategory
-from deepchecks.checks.distribution.trust_score import TrustScore
-from deepchecks.checks.distribution.preprocessing import ScaledNumerics
-from deepchecks.checks.distribution.plot import get_density
+from deepchecks.utils.distribution.trust_score import TrustScore
+from deepchecks.utils.distribution.preprocessing import ScaledNumerics
+from deepchecks.utils.distribution.plot import get_density
 from deepchecks.utils.metrics import task_type_check, ModelType
 from deepchecks.utils.strings import format_percent
 from deepchecks.utils.validation import validate_model
