+++ deepchecks/utils/distribution/preprocessing.py
@@ -21,7 +21,7 @@ from sklearn.base import TransformerMixin, BaseEstimator
 from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import MinMaxScaler
 from deepchecks.utils.typing import Hashable
-from deepchecks.checks.distribution.rare_category_encoder import RareCategoryEncoder
+from deepchecks.utils.distribution.rare_category_encoder import RareCategoryEncoder
 
 
 __all__ = ['ScaledNumerics', 'preprocess_2_cat_cols_to_same_bins']
